fixed bug, where forward_euler_step_size=0

+++ controller/src/angular_vel_to_pwm_node.cpp
@@ -21,7 +21,7 @@ int main(int argc, char** argv)
 
     ros::Publisher pwm_pub = w2pwm_nh.advertise<asl_gremlin_msgs::MotorPwm>
                                                         (pwm_pub_topic,20);
-    int rate = 10;
+    double rate = 10.0;
     if (!w2pwm_nh.getParam(ros::this_node::getNamespace()+"/sim/rate", rate))
     {
         ROS_WARN("Unable access parameter $robot_name/sim/rate, setting rate as 10Hz");
