+++ controller/src/backstepping_controller_node.cpp
@@ -40,7 +40,7 @@ int main(int argc, char** argv)
     ros::Publisher ang_vel_cmd = ctrl_nh.advertise<MotorAngVel>
                                                     (ang_vel_topic, 20);
  
-    int rate = 10;
+    double rate = 10.0;
     if (!ctrl_nh.getParam(ros::this_node::getNamespace()+"/sim/rate", rate))
     {
         ROS_WARN("Unable access parameter $robot_name/sim/rate, setting rate as 10Hz");
