+++ state_feedback/src/feedback_selected.cpp
@@ -8,7 +8,7 @@ int main(int argc, char** argv)
 
     state_feedback::FeedbackSelected<3> feedback_method(feedback_nh);
 
-    int rate = 10;
+    double rate = 10.0;
     if (!feedback_nh.getParam(ros::this_node::getNamespace()+"/sim/rate", rate))
     {
         ROS_WARN("Unable access parameter $robot_name/sim/rate, setting rate as 10Hz");
